test(task): fold new repair-args cases into existing test file
Address review feedback: origin/main already had
packages/coding-agent/test/tools/task-repair-args.test.ts covering
the same repairDoubleEncodedJsonString contracts.

- Removed duplicate packages/coding-agent/test/task/repair-args.test.ts
- Added only new cases to the existing file:
  - backslash-backslash decode
  - batch form context repair
  - per-task-item repair
  - null/undefined defensive handling
  - partially streamed params (missing fields)
